requirements.txt
python-dotenv==1.0.1
paramiko==3.4.0
tqdm==4.66.1
fpdf==1.7.2
sqlite3-cli==0.1.0